server/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.1.0
fastapi==0.116.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.16.0
h5py==3.12.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==20.9
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.47.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.12.0
tensorflow_intel==2.18.0
tensorflowjs==3.18.0
termcolor==2.5.0
torch==2.5.1
tornado==6.4.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.43.0
wrapt==1.16.0
